+++ RELEASE_NOTES.md
@@ -3,3 +3,6 @@
 - **Virtual Pinball Spreadsheet**: Added missing "open" link for the table selection.
 - **Virtual Pinball Spreadsheet**: Added update date to table selection.
 - **Preferences**: Added **VPin Mame** section.
+- **Popper Media View**: Fixed missing Apron/Full DMD media view.
+- **Popper Media View**: Fixed media disposing error that happened for some videos and did freeze the whole UI.
+-